tests/wpt/tests/css/cssom/cssstyledeclaration-csstext-final-delimiter.html

<!DOCTYPE html>
<html>
  <head>
    <title>CSSOM - CSSStyleDeclaration - Text - Serialization - Delimiters</title>
    <link rel="author" title="Glenn Adams" href="mailto:glenn@skynav.com"/>
    <link rel="help" href="https://drafts.csswg.org/cssom/#the-cssstyledeclaration-interface"/>
    <meta name="flags" content="dom"/>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="box"></div>
    <script>
      var style = document.getElementById('box').style;

      test(function(){
        style.cssText = "";
        assert_equals(style.cssText, "");
      }, 'inline style - text - delimiters - zero declarations');

      test(function(){
        style.cssText = "left: 10px";
        assert_equals(style.cssText, "left: 10px;");
      }, 'inline style - text - delimiters - one declaration');

      test(function(){
        style.cssText = "left: 10px; right: 20px";
        assert_equals(style.cssText, "left: 10px; right: 20px;");
      }, 'inline style - text - delimiters - two declarations');
    </script>
  </body>
</html>
